+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+# [1.6.0](https://github.com/RouHim/schweinehund/compare/v1.5.0...v1.6.0) (2026-02-11)
+
+
+### Features
+
+* **db:** add get_tasks_for_month calendar query function ([b08421a](https://github.com/RouHim/schweinehund/commit/b08421ae37c729235f34e6578116eb0aa34511a8))
+* **routes:** add GET /api/tasks/calendar endpoint ([75eaaf2](https://github.com/RouHim/schweinehund/commit/75eaaf24006777d968818550c25f204da0969cca))
+* **ui:** add calendar tab and CSS grid layout ([c1fac5c](https://github.com/RouHim/schweinehund/commit/c1fac5cf47f9d7ca98dd764c2ce2674ddc952aaf))
+* **ui:** implement calendar JavaScript logic ([72f8b81](https://github.com/RouHim/schweinehund/commit/72f8b815177011394316d8caf1ba8d282776974b))
+
+# [1.5.0](https://github.com/RouHim/schweinehund/compare/v1.4.0...v1.5.0) (2026-02-11)
+
+
+### Features
+
+* **api:** accept user-provided start_date in task create/update endpoints ([3fabc94](https://github.com/RouHim/schweinehund/commit/3fabc94752066fc47076006558306d682743731a))
+* **db:** add start_date filtering to hide future-dated tasks from today view ([bd491ae](https://github.com/RouHim/schweinehund/commit/bd491ae63a99ef08622d9562473c1d264ebd1ae7))
+* **ui:** add start date picker to task modal with future-date badge ([3467501](https://github.com/RouHim/schweinehund/commit/3467501ed9767c08d1f02e69c27a740e920ef4e3))
+
 # [1.4.0](https://github.com/RouHim/schweinehund/compare/v1.3.0...v1.4.0) (2026-02-11)